Game mechanics done, almost two levels complete
At this point in the project, I realized that the game was growing
beyond the point of "no risks", so I put it in GitHub.

There are almost two levels at this point with transitions between
levels, fluid animations for the spawning and removing of elements on
the screen, and an infinitely expandable, easily programmable recipe
file. This game is going places!